vision/students_access/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
imutils==0.5.4
inflection==0.5.1
numpy==2.1.3
opencv-contrib-python==4.10.0.84
packaging==24.2
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
whitenoise==6.8.2
